Rename Services & Remove Response Handling From Service
